Revert "Revert "Added basic music integration""
This reverts commit 58f7fafc8637fb4e0f4a24b1ffc191df9e4384c2.

+++ src/main/java/me/imlukas/slashcommands/commands/xp/listener/UserMessageListener.java
@@ -15,6 +15,7 @@ import java.util.concurrent.ThreadLocalRandom;
 
 public class UserMessageListener extends ListenerAdapter {
 
+    // TODO: RE-WRITE XP SYSTEM
     private final Bot main;
     private final Random random = ThreadLocalRandom.current();
 
@@ -36,11 +37,10 @@ public class UserMessageListener extends ListenerAdapter {
 
         Guild guild = event.getGuild();
 
-        main.getSqlHandler().fetch(xpData, event.getGuild(), user).thenAccept((oldXp) -> {
+        main.getSqlHandler().fetch(xpData, event.getGuild(), user).thenAccept(( oldXp) -> {
             main.getSqlHandler().addXp(xpData, xp, guild, user);
-
-            int newXp = oldXp + xp;
-            int currentLevel = XpUtil.getLevelFromXp(oldXp);
+            int newXp = (int) oldXp + xp;
+            int currentLevel = XpUtil.getLevelFromXp((int) oldXp);
             int nextLevelXpNeeded = XpUtil.getXpToLevel(currentLevel + 1);
             int nextLevelXp = nextLevelXpNeeded - newXp;
 
@@ -54,10 +54,6 @@ public class UserMessageListener extends ListenerAdapter {
                 event.getChannel().sendMessage("Congrats " + user.getAsMention() + " you leveled up to level " + (currentLevel + 1)).queue();
                 return;
             }
-
-
-
-
 
         });
